requirements.txt
pandas>=2.0
requests>=2.28
beautifulsoup4>=4.12
lxml>=5.0
html5lib>=1.1
feedparser>=6.0
finance-datareader>=0.9.50
python-dotenv>=1.0
